requirements/common.txt
Flask==2.0.2
mediacloud==3.12.5
requests==2.26.0
raven[flask]==6.10.0
redis==3.5.3
gunicorn==20.1.0
dogpile.cache==0.7.1  # upgrading dogpile.cache caused problems - leave it on this version
python-dotenv==0.19.1
python-dateutil==2.8.2
dateparser==1.1.0
